MNT: Move bluesky.tests.test_new_examples.Status to bluesky.examples.NullStatus
